requirements.txt
asgiref==3.3.1
Django==3.1.6
django-imap-backend==0.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
Pillow==8.1.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
requests==2.25.1
gunicorn
psycopg2==2.8.6
psycopg2-binary==2.8.6